Updated to use MatchIdentifier struct
