+++ Cdms.Model/MatchIdentifier.cs
@@ -1,0 +1,49 @@
+using Cdms.Model.Ipaffs;
+
+namespace Cdms.Model;
+
+public struct MatchIdentifier(int identifier)
+{
+    public int Identifier { get; private set; } = identifier;
+
+    public static MatchIdentifier FromNotification(string reference)
+    {
+        if (reference == null)
+        {
+            throw new ArgumentNullException(nameof(reference));
+        }
+
+        var parts = reference.Split(".");
+        int identifier;
+        if (char.IsDigit(parts[3].Last()))
+        {
+            identifier = int.Parse(parts[3]);
+        }
+        else
+        {
+            identifier = int.Parse(parts[3].Remove(parts[3].Length - 1));
+        }
+
+        return new MatchIdentifier(identifier);
+    }
+
+    public static MatchIdentifier FromCds(string reference)
+    {
+        // var reference = document.DocumentReference;
+
+        int identifier;
+        var parts = reference.Split(".");
+
+        var identifierString = parts.Last();
+        if (char.IsDigit(identifierString.Last()))
+        {
+            identifier = int.Parse(identifierString);
+        }
+        else
+        {
+            identifier = int.Parse(identifierString.Remove(identifierString.Length - 1));
+        }
+
+        return new MatchIdentifier(identifier);
+    }
+}
